requirements.txt
# API & HTTP
requests>=2.31.0
httpx>=0.25.0

# Scraping
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data & Storage
sqlalchemy>=2.0.0
pandas>=2.0.0

# Scheduling (optional - can use cron instead)
schedule>=1.2.0

# API server
fastapi>=0.109.0
uvicorn>=0.27.0

# Environment
python-dotenv>=1.0.0
